While working on OPENTHREAD-1124 I discovered that there were some descrepencies between GitHub OpenThread and our stash OpenThread so I compared the devel/21q4 branch to the master branch and looked at all the diffs.  I expected diffs associated with multipan and coprocessor cli support so I ignored those differences.  However, there were some differences that need to be fixed and I'm addressing them in this commit.  These were most likely introduced by bad merges.
This is the URL I used to compare the branches.  You can easily do this using the compare option in stash.

https://stash.silabs.com/projects/WMN_STACKS/repos/openthread/compare/diff?targetBranch=refs%2Fheads%2Fmaster&sourceBranch=refs%2Fheads%2Fdevelop%2F21q4&targetRepoId=2181

Suvesh also informed me that is can be done this way:

I do this on my branch like so (finds the common parent with GitHub master to compare to, in case master is ahead since we last merged):

git merge-base $(git rev-parse --abbrev-ref HEAD) origin/master | xargs -I{} git diff {}