OPENTHREAD-2721 - CLI compilation issues with ot_ncp and ot_rcp apps
The issue was introduced by the recent update to openthread.

The otCliCommand data struct was changed which means we need to modify
all of our custom CLI function definitions.

The change was such that the CLI commands can now return an error.
Because of this we no longer need to separately store the user command
error using the Interpreter::SetUserCommandError function.  Instead
we can just return the error.